chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-success-fail-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
FAIL toSuccess.import.querySelector('h1').innerHTML should be Hello. Threw exception TypeError: Cannot read property 'querySelector' of undefined
FAIL toSuccess.import.querySelector('html') should be non-null. Threw exception TypeError: Cannot read property 'querySelector' of undefined
FAIL toFail.import should be null (of type object). Was undefined (of type undefined).
FAIL notImport.import should be null (of type object). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
